Ajout réseau unique pour l'appli

+++ configEnvironnement.ps1
@@ -1,4 +1,6 @@
-﻿# Détection du proxy sur le système pour le configurer sur Docker
+﻿#Requires -RunAsAdministrator
+
+# Détection du proxy sur le système pour le configurer sur Docker
 
  Write-host ""
  Write-host "============================================"
@@ -56,6 +58,45 @@ if ($adresseProxy -ilike "*=*")
 
 
 
+# Configuration du réseau pour l'application
+
+ Write-host ""
+ Write-host "=========================================="
+ Write-host "Configuration du réseau pour l'application"
+ Write-host "=========================================="
+ Write-host ""
+
+ Write-Host "L'application utilise le réseau 192.168.97.0/24 avec une passerelle de 192.168.97.1." 
+ Write-Host "Vous pouvez changer ces paramètres si vous pensez qu'il peut y avoir un conflit d'adresses IP avec votre réseau existant et si vous savez ce que vous faites"
+ Write-Host ""
+ 
+ $accord=Read-Host "Veuillez confirmer que vous désirez modifier le réseau (répondre par oui pour confirmer ou par n'importe quel autre caractère pour maintenir le réseau actuel)"
+     if ($accord -eq "oui") 
+      {
+      Read-Host "Saisissez le réseau sous la forme adresseIP/CIDR" $NET_ECB
+      Read-Host "Saisissez l'adresse IP de la passerelle" $GW_ECB
+      }
+     else {
+      $NET_ECB = "192.168.97.0/24"
+      $GW_ECB = "192.168.97.1"
+      }
+
+
+ if ((docker network ls) | Select-String bridge_e-combox) {
+   Write-host "    --> Le réseau existe, il faut le supprimer avant de le créer à nouveau." -Fore blue
+   docker stop $(docker ps -q)
+   docker network rm bridge_e-combox
+   docker network create --subnet $NET_ECB --gateway=$GW_ECB bridge_e-combox
+   }
+else {
+   Write-host "    --> Le réseau n'existe pas, il faut le créer." -Fore blue
+   docker network create --subnet $NET_ECB --gateway=$GW_ECB bridge_e-combox
+}
+
+Write-Host "L'application utilise maintenant le réseau $NET_ECB avec une passerelle de $GW_ECB."
+
+Read-Host "Appuyez sur la touche Entrée pour continuer"
+
 #Détection et configuration de l'adresse IP
 
  Write-host ""
@@ -86,7 +127,7 @@ If ($docker_ip_host -eq $adressesIPvalides) {
  write-host ""
  Write-host "Le système a détecté que vous utilisez cette adresse IP : $docker_ip_host et que vous ne disposez pas d'autres adresses IP valides susceptibles d'être configurées avec e-comBox."
  Write-Host ""
- Read-Host "Appuyez sur la touche Entrée pour fermer ce programme"
+ #Read-Host "Appuyez sur la touche Entrée pour fermer ce programme"
  }
  else {
   Write-host ""
@@ -110,7 +151,7 @@ If ($docker_ip_host -eq $adressesIPvalides) {
 
      Write-host ""
 
-     $confirmation=Read-Host "Veuillez confirmer que vous désirez utiliser la nouvelle adresse IP $adresseIP pour rendre accessible vos sites dans e-comBox (répondre par oui pour confirmer ou par n'importe quel autre caractère pour maintenir l'adresse IP actuelle)"
+     $confirmation=Read-Host "Veuillez confirmer que vous désirez utiliser la nouvelle adresse IP $adresseIP pour rendre accessible vos sites dans e-comBox (répondre par oui pour confirmer ou par n'importe quel autre caractère pour maintenir l'adresse IP actuelle)."
      if ($confirmation -eq "oui") 
       {
       $docker_ip_host=$adresseIP
@@ -129,24 +170,71 @@ URL_UTILE=$docker_ip_host
 
      Write-host ""
      Write-host "Le système va maintenant configurer e-comBox avec l'adresse IP : $docker_ip_host et lancer l'application dans votre navigateur par défaut."
-     Write-host ""
-     sleep 5
-
-   # Redémarrage de Portainer
-   Set-Location -Path $env:USERPROFILE\e-comBox_portainer\
-   docker-compose down
-   docker-compose up -d
-
-   # Redémarrage de l'application
-   docker rm -f e-combox
-   docker run -dit --name e-combox -v ecombox_data:/usr/local/apache2/htdocs/ --restart always -p 8888:80 aporaf/e-combox:1.0
-   Start-Process "C:\Program Files\e-comBox\e-comBox.url"
-   
+     Write-host ""    
+     #sleep 5   
    }
       else {
        Write-host ""
        Write-Host "L'application e-comBox continuera à utiliser l'adresse ip $docker_ip_host."
-       sleep 5
+      
+       #sleep 5
      }
-  #}
 }
+
+# Détection et configuration d'un éventuel proxy pour Git
+
+Set-Location -Path $env:USERPROFILE
+
+$reg = "HKCU:\Software\Microsoft\Windows\CurrentVersion\Internet Settings"
+$settings = Get-ItemProperty -Path $reg
+
+if ($settings.ProxyEnable -eq 1) {
+    $adresseProxy = $settings.ProxyServer
+    if ($adresseProxy -ilike "*=*")
+        {
+            $adresseProxy = $adresseProxy -replace "=","://" -split(';') | Select-Object -First 1
+        }
+
+        else
+        {
+            $adresseProxy = "http://" + $adresseProxy
+        }
+    #Write-Host "l'adresse du proxy est $adresseProxy"
+    git config --global http.proxy $adresseProxy
+    }
+    else {
+      git config --global --unset http.proxy
+      }
+
+
+# Récupération de portainer
+
+$Path="$env:USERPROFILE\e-comBox_portainer\"
+$TestPath=Test-Path $Path
+
+If ($TestPath -eq $False) {
+    # Installation de Portainer sur Git
+    Write-host "    --> Portainer n'est pas installé, il faut l'installer." -Fore Red
+    git clone https://github.com/siollb/e-comBox_portainer.git 2>$null
+    }
+    else {
+      # Arrêt de Portainer
+      Write-host "    --> Portainer est démarré, il faut le stopper et le supprimer pour le réinstaller." -Fore blue
+      Set-Location -Path $env:USERPROFILE\e-comBox_portainer\
+      docker-compose down
+      Set-Location -Path $env:USERPROFILE\
+      Remove-Item "e-comBox_portainer" -Recurse    
+      git clone https://github.com/siollb/e-comBox_portainer.git 2>$null       
+      } 
+
+# Démarrage de Portainer
+   Set-Location -Path $env:USERPROFILE\e-comBox_portainer\
+   docker-compose up --build -d
+
+# Redémarrage de l'application
+   docker rm -f e-combox
+   docker pull aporaf/e-combox:1.0
+   docker run -dit --name e-combox -v ecombox_data:/usr/local/apache2/htdocs/ --restart always -p 8888:80 --network bridge_e-combox aporaf/e-combox:1.0
+   Start-Process "C:\Program Files\e-comBox\e-comBox.url"
+
+Read-Host "Appuyez sur la touche Entrée pour fermer ce programme"
